+++ Import/import.cpp
@@ -17,17 +17,14 @@
 #include <utility>
 #include <vector>
 
-#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "stb_image_write.h"
-
-#define GLAD_GLES2_IMPLEMENTATION
-#include "gles2.h"
-
 #include "ToolKit.h"
 #include "Util.h"
 #include "glm/gtc/matrix_transform.hpp"
 #include "glm/gtx/quaternion.hpp"
 #include "rapidxml.hpp"
+
+#define STB_IMAGE_WRITE_IMPLEMENTATION
+#include "stb_image_write.h"
 
 using std::cout;
 using std::endl;
