+++ ToolKit/Renderer.h
@@ -188,12 +188,9 @@ namespace ToolKit
     typedef struct RHIConstants
     {
       static constexpr ubyte textureSlotCount       = 32;
-
       static constexpr size_t maxLightsPerObject    = 16;
-
       static constexpr int shadowAtlasSlot          = 8;
       static constexpr int g_shadowAtlasTextureSize = 4096;
-
       static constexpr int specularIBLLods          = 5;
     } m_rhiSettings;
